Document and split inst match trie to new class.
